+++ picklepipe/marshalpipe.py
@@ -23,12 +23,12 @@ class _MarshalSerializer(object):
 
 class MarshalPipe(BaseSerializingPipe):
     """ Wraps an already connected socket and uses that
-    socket as a interface to send marshalled objects to a peer.
+    socket as a interface to send marshaled objects to a peer.
     Can be used to pickle not only single objects but also
     to pickle objects in a stream-able fashion. """
     def __init__(self, sock, version=None):
         """
-        Creates a :class:`picklepipe.PicklePipe` instance wrapping
+        Creates a :class:`picklepipe.MarshalPipe` instance wrapping
         a given socket.
 
         :param sock: Socket to wrap.
@@ -58,7 +58,7 @@ class MarshalPipe(BaseSerializingPipe):
         super(MarshalPipe, self).send_object(obj)
 
     def recv_object(self, timeout=None):
-        """ Receives a marshalled object from the peer.
+        """ Receives a marshaled object from the peer.
 
         :param float timeout: Number of seconds to wait before timing out.
         :return: Pickled object or None if timed out.
@@ -66,6 +66,10 @@ class MarshalPipe(BaseSerializingPipe):
         """
         self._recv_version()
         return super(MarshalPipe, self).recv_object(timeout)
+
+    def fileno(self):
+        self._recv_version()
+        return super(MarshalPipe, self).fileno()
 
     def _send_version(self):
         if not self._version_sent: