+++ CHANGELOG.rst
@@ -8,3 +8,4 @@ Release 0.0.1 (Development)
 * Added support for creating own serialization pipes with :class:`picklepipe.BaseSerializationPipe`.
 * Added support for the ``marshal`` object serializer with :class:`picklepipe.MarshalPipe`
 * :class:`picklepipe.PicklePipe` now uses ``cPickle`` module if available.
+* Make all pipe types selectable using selectors.