+++ CHANGELOG.rst
@@ -1,8 +1,13 @@
 Changelog
 =========
 
-Release 0.0.1 (Development)
+Release 1.0.1 (Development)
 ---------------------------
+
+* (Add summary of changes here, along with author.)
+
+Release 1.0.0 (December 28, 2016)
+---------------------------------
 * Create initial implementation of :class:`picklepipe.PicklePipe`.
 * Create :meth:`picklepipe.make_pipe_pair` to create a connected pair of :class:`picklepipe.PicklePipe` instances.
 * Added support for creating own serialization pipes with :class:`picklepipe.BaseSerializationPipe`.